Perform housekeeping
