+++ Microsoft.Azure.WebJobs.Extensions.OpenApi.sln
@@ -28,9 +28,11 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "samples", "samples", "{66D8
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.Azure.Functions.Worker.Extensions.OpenApi.FunctionApp.V3Net5", "samples\Microsoft.Azure.Functions.Worker.Extensions.OpenApi.FunctionApp.V3Net5\Microsoft.Azure.Functions.Worker.Extensions.OpenApi.FunctionApp.V3Net5.csproj", "{BDE6F761-880E-47C3-B40C-956F470E3D10}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.Azure.WebJobs.Extensions.OpenApi.FunctionApp.Interfaces", "samples\Microsoft.Azure.WebJobs.Extensions.OpenApi.FunctionApp.Interfaces\Microsoft.Azure.WebJobs.Extensions.OpenApi.FunctionApp.Interfaces.csproj", "{F1438954-24F5-4D41-AFEA-46D9E1E74C72}"
+EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.Azure.WebJobs.Extensions.OpenApi.FunctionApp.Models", "samples\Microsoft.Azure.WebJobs.Extensions.OpenApi.FunctionApp.Models\Microsoft.Azure.WebJobs.Extensions.OpenApi.FunctionApp.Models.csproj", "{6ABA1666-BC42-487F-B156-3DC79347BD6B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.Azure.WebJobs.Extensions.OpenApi.FunctionApp.Interfaces", "samples\Microsoft.Azure.WebJobs.Extensions.OpenApi.FunctionApp.Interfaces\Microsoft.Azure.WebJobs.Extensions.OpenApi.FunctionApp.Interfaces.csproj", "{F1438954-24F5-4D41-AFEA-46D9E1E74C72}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.Azure.WebJobs.Extensions.OpenApi.FunctionApp.Models.PetStore", "samples\Microsoft.Azure.WebJobs.Extensions.OpenApi.FunctionApp.Models.PetStore\Microsoft.Azure.WebJobs.Extensions.OpenApi.FunctionApp.Models.PetStore.csproj", "{0FBF046E-7B17-410C-8219-8381D07726DF}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.Azure.WebJobs.Extensions.OpenApi.FunctionApp.Services", "samples\Microsoft.Azure.WebJobs.Extensions.OpenApi.FunctionApp.Services\Microsoft.Azure.WebJobs.Extensions.OpenApi.FunctionApp.Services.csproj", "{B7CF1C77-FE6F-4586-93FE-801671A5EAFD}"
 EndProject
@@ -90,10 +92,18 @@ Global
 		{BDE6F761-880E-47C3-B40C-956F470E3D10}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BDE6F761-880E-47C3-B40C-956F470E3D10}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{BDE6F761-880E-47C3-B40C-956F470E3D10}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F1438954-24F5-4D41-AFEA-46D9E1E74C72}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F1438954-24F5-4D41-AFEA-46D9E1E74C72}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F1438954-24F5-4D41-AFEA-46D9E1E74C72}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F1438954-24F5-4D41-AFEA-46D9E1E74C72}.Release|Any CPU.Build.0 = Release|Any CPU
 		{6ABA1666-BC42-487F-B156-3DC79347BD6B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{6ABA1666-BC42-487F-B156-3DC79347BD6B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6ABA1666-BC42-487F-B156-3DC79347BD6B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6ABA1666-BC42-487F-B156-3DC79347BD6B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0FBF046E-7B17-410C-8219-8381D07726DF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0FBF046E-7B17-410C-8219-8381D07726DF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0FBF046E-7B17-410C-8219-8381D07726DF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0FBF046E-7B17-410C-8219-8381D07726DF}.Release|Any CPU.Build.0 = Release|Any CPU
 		{B7CF1C77-FE6F-4586-93FE-801671A5EAFD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{B7CF1C77-FE6F-4586-93FE-801671A5EAFD}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B7CF1C77-FE6F-4586-93FE-801671A5EAFD}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -146,10 +156,6 @@ Global
 		{65AED6F0-D27E-4D6E-BA42-AD6EE1338082}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{65AED6F0-D27E-4D6E-BA42-AD6EE1338082}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{65AED6F0-D27E-4D6E-BA42-AD6EE1338082}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F1438954-24F5-4D41-AFEA-46D9E1E74C72}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F1438954-24F5-4D41-AFEA-46D9E1E74C72}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F1438954-24F5-4D41-AFEA-46D9E1E74C72}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F1438954-24F5-4D41-AFEA-46D9E1E74C72}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -162,7 +168,9 @@ Global
 		{7AD094AA-01D6-4644-B9B6-E555A029D6D7} = {810145E0-41CF-4E24-BD9C-E7517498BA29}
 		{66D8DEA8-B477-497F-95BB-E8F9A5BAC352} = {810145E0-41CF-4E24-BD9C-E7517498BA29}
 		{BDE6F761-880E-47C3-B40C-956F470E3D10} = {66D8DEA8-B477-497F-95BB-E8F9A5BAC352}
+		{F1438954-24F5-4D41-AFEA-46D9E1E74C72} = {66D8DEA8-B477-497F-95BB-E8F9A5BAC352}
 		{6ABA1666-BC42-487F-B156-3DC79347BD6B} = {66D8DEA8-B477-497F-95BB-E8F9A5BAC352}
+		{0FBF046E-7B17-410C-8219-8381D07726DF} = {66D8DEA8-B477-497F-95BB-E8F9A5BAC352}
 		{B7CF1C77-FE6F-4586-93FE-801671A5EAFD} = {66D8DEA8-B477-497F-95BB-E8F9A5BAC352}
 		{98223494-E0A8-4C07-A38F-84CA2060031D} = {66D8DEA8-B477-497F-95BB-E8F9A5BAC352}
 		{848EDB8F-9C45-462B-A544-EA237774F3BA} = {66D8DEA8-B477-497F-95BB-E8F9A5BAC352}
@@ -176,7 +184,6 @@ Global
 		{968CD847-80D0-4740-92D5-F5553461EC04} = {8B62E3FB-9062-4716-803A-1FA51FCE68BC}
 		{D787694D-B87B-4E11-9455-B5BF9306D8FC} = {8B62E3FB-9062-4716-803A-1FA51FCE68BC}
 		{65AED6F0-D27E-4D6E-BA42-AD6EE1338082} = {8B62E3FB-9062-4716-803A-1FA51FCE68BC}
-		{F1438954-24F5-4D41-AFEA-46D9E1E74C72} = {66D8DEA8-B477-497F-95BB-E8F9A5BAC352}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {70FEC644-BB3C-4441-AF91-DC694803C8F2}
