+++ samples/Microsoft.Azure.Functions.Worker.Extensions.OpenApi.FunctionApp.V3Net5/Program.cs
@@ -1,4 +1,4 @@
-using Microsoft.Azure.Functions.Worker.Extensions.OpenApi.Extensions;
+using Microsoft.Azure.Functions.Worker.Extensions.OpenApi.FunctionApp.V3Net5.Extensions;
 using Microsoft.Extensions.Hosting;
 
 namespace Microsoft.Azure.Functions.Worker.Extensions.OpenApi.FunctionApp.V3Net5
