+++ samples/Microsoft.Azure.WebJobs.Extensions.OpenApi.FunctionApp.Models.PetStore/Category.cs
@@ -1,12 +1,8 @@
-using Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Attributes;
-using Microsoft.Azure.WebJobs.Extensions.OpenApi.FunctionApp.Models.Examples;
-
-namespace Microsoft.Azure.WebJobs.Extensions.OpenApi.FunctionApp.Models
+namespace Microsoft.Azure.WebJobs.Extensions.OpenApi.FunctionApp.Models.PetStore
 {
     /// <summary>
     /// This represents the model entity for category of Swagger Pet Store.
     /// </summary>
-    [OpenApiExample(typeof(CategoryExample))]
     public class Category
     {
         /// <summary>
