+++ samples/Microsoft.Azure.WebJobs.Extensions.OpenApi.FunctionApp.Models.PetStore/Pet.cs
@@ -1,10 +1,8 @@
 using System.Collections.Generic;
 
-using Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Attributes;
-
 using Newtonsoft.Json;
 
-namespace Microsoft.Azure.WebJobs.Extensions.OpenApi.FunctionApp.Models
+namespace Microsoft.Azure.WebJobs.Extensions.OpenApi.FunctionApp.Models.PetStore
 {
     /// <summary>
     /// This represents the model entity for pet of Swagger Pet Store.
@@ -41,7 +39,6 @@ namespace Microsoft.Azure.WebJobs.Extensions.OpenApi.FunctionApp.Models
         /// <summary>
         /// Gets or sets the <see cref="PetStatus"/> value.
         /// </summary>
-        [OpenApiProperty(Description = "pet status in the store")]
         public PetStatus? Status { get; set; }
     }
 }
